[refactor] Move Destinations to ui/utils
