+++ app/src/main/java/com/aruzeta/tmdb/ui/utils/Destination.kt
@@ -1,4 +1,4 @@
-package com.aruzeta.tmdb.ui
+package com.aruzeta.tmdb.ui.utils
 
 import androidx.navigation.NamedNavArgument
 
@@ -6,5 +6,5 @@ sealed class Destination(
     val route: String,
     val arguments: List<NamedNavArgument> = emptyList(),
 ) {
-    object FirstScreen: Destination("first")
+    object FirstScreen : Destination("first")
 }
